refactor json
- Split out code in many submodules. Following repo conventions.
- The API is backwards compatible in this change.
