+++ src/json/json.ts
@@ -1,36 +1,15 @@
 import { z } from "zod";
+import type { JsonArray } from "./array.js";
+import type { JsonObject } from "./object.js";
+import { type JsonPrimitive, primitive } from "./primitive.js";
 
-const primitiveSchema = z.union([
-  z.string(),
-  z.number(),
-  z.boolean(),
-  z.null(),
-]);
-export type JsonPrimitive = z.infer<typeof primitiveSchema>;
-export function primitive() {
-  return primitiveSchema;
-}
+const schema: z.ZodType<Json> = z.lazy(() =>
+  z.union([primitive(), z.array(schema), z.record(z.string(), schema)]),
+);
 
-export type JsonObject = { [key: string]: Json };
-export type JsonArray = Array<Json>;
 export type Json = JsonPrimitive | JsonObject | JsonArray;
 
-const jsonSchema: z.ZodType<Json> = z.lazy(() =>
-  z.union([primitive(), z.array(jsonSchema), z.record(z.string(), jsonSchema)]),
-);
-export function json() {
-  return jsonSchema;
-}
-
-const objectSchema = z.record(z.string(), jsonSchema);
-export function object() {
-  return objectSchema;
-}
-
-const arraySchema = z.array(jsonSchema);
-export function array() {
-  return arraySchema;
-}
+export const json = () => schema;
 
 // TODO: codec stringToJson.
 export const stringifiedJsonSchema: z.ZodType<Json, string> = z
