+++ src/json/types.ts
@@ -1,1 +1,4 @@
-export type { Json, JsonArray, JsonObject, JsonPrimitive } from "./json.js";
+export type { JsonArray } from "./array.js";
+export type { Json } from "./json.js";
+export type { JsonObject } from "./object.js";
+export type { JsonPrimitive } from "./primitive.js";
